add number float

+++ src/main/java/org/example/my/task/generatenumberfx/RandomNumberGenerator.java
@@ -1,5 +1,5 @@
 package org.example.my.task.generatenumberfx;
 
 public interface RandomNumberGenerator {
-    long generate();
+    float generate();
 }
